backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
python-pptx==0.6.23
groq==0.4.2
python-dotenv==1.0.0
